crewnews/pyproject.toml
[project]
name = "crewnews"
version = "0.1.0"
description = "crewnews using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "arize-phoenix>=8.8.0", 
    "arize-phoenix-client>=1.0.3",
    "arize-phoenix-evals>=0.20.3",
    "arize-phoenix-otel>=0.8.0",
    "opentelemetry-sdk>=1.15.0", 
    "opentelemetry-exporter-otlp>=1.15.0",
    "openinference-instrumentation-crewai>=0.1.0"
]

[project.scripts]
crewnews = "crewnews.main:run"
run_crew = "crewnews.main:run"
train = "crewnews.main:train"
replay = "crewnews.main:replay"
test = "crewnews.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
